Use task configuration avoidance in gradle initialization
Gradle task creation can be heavyweight when hundreds or thousands of
instances of tasks exist. This is especially true in large multi project
builds. Forbidden apis currently always creates a task per source set,
as well as the alias `forbiddenApis` task. Gradle provides a way to
lazily construct the task, and only configure when the task is actually
needed for a given task graph.

This commit switches the construction of the forbidden apis tasks to use
task registration, when it is available (based on gradle version).